refactor: 拆分 config 扩展
