+++ src/Extensions/Components/BootstrapBlazor.Dock/wwwroot/js/golden-layout-extensions.js
@@ -106,6 +106,22 @@ const getConfig = option => {
     }
 }
 
+const getAllItemsByType = (type, parent) => {
+    const items = []
+
+    parent.content.forEach(v => {
+        if (v.type === type) {
+            v.parent = parent;
+            items.push(v)
+        }
+
+        if (v.content != null) {
+            items.push.apply(items, getAllItemsByType(type, v))
+        }
+    })
+    return items
+}
+
 const resetComponentId = (config, option) => {
     // 本地配置
     const localComponents = getAllItemsByType('component', config.root)
@@ -182,22 +198,6 @@ const resetComponentId = (config, option) => {
             show: stack.hasHeaders
         }
     });
-}
-
-const getAllItemsByType = (type, parent) => {
-    const items = []
-
-    parent.content.forEach(v => {
-        if (v.type === type) {
-            v.parent = parent;
-            items.push(v)
-        }
-
-        if (v.content != null) {
-            items.push.apply(items, getAllItemsByType(type, v))
-        }
-    })
-    return items
 }
 
 const createItem = item => ({
